+++ database/migrations/2018_07_09_203244_PagosMensualesArrendatarios.php
@@ -33,11 +33,11 @@ class PagosMensualesArrendatarios extends Migration
             $table->double('pago_a_arrendatario_moneda', 20, 8)->nullable();;
             $table->double('pago_a_rentas_moneda', 20, 8)->nullable();;
             $table->double('valor_a_pagar_moneda', 20, 8)->nullable();;
-            $table->integer('subtotal_entrada')->nullable();;
-            $table->integer('subtotal_salida')->nullable();;
-            $table->integer('pago_a_arrendatario')->nullable();;
-            $table->integer('pago_a_rentas')->nullable();;
-            $table->integer('valor_a_pagar')->nullable();;
+            $table->double('subtotal_entrada', 20,2)->nullable();;
+            $table->double('subtotal_salida', 20,2)->nullable();;
+            $table->double('pago_a_arrendatario', 20,2)->nullable();;
+            $table->double('pago_a_rentas', 20,2)->nullable();;
+            $table->double('valor_a_pagar', 20,2)->nullable();;
             $table->integer('id_creador');
             $table->integer('id_modificador');
             $table->integer('id_estado');
