+++ database/migrations/2018_08_04_171413_DetalleSolicitudServicio.php
@@ -27,7 +27,7 @@ class DetalleSolicitudServicio extends Migration
             $table->date('fecha_uf');
             $table->double('valor_uf',20,8);
             $table->double('valor_en_uf',20,8);
-            $table->integer('valor_en_pesos');
+            $table->double('valor_en_pesos',20,2);
             $table->integer('cantidad');
             $table->double('subtotal_uf',20,8);
             $table->integer('subtotal_pesos');
