+++ database/migrations/2018_08_05_020226_SolicitudServicios_ARR.php
@@ -27,7 +27,7 @@ class SolicitudServiciosARR extends Migration
             $table->date('fecha_uf');
             $table->double('valor_uf',20,8);
             $table->double('valor_en_uf',20,8);
-            $table->integer('valor_en_pesos');
+            $table->double('valor_en_pesos',20,2);
             $table->integer('id_estado')->nullable();
             $table->timestamps();
         });
